[dx2d] 클래스 괄호 없던 오류 수정 / Camera cpp h 유니코드 변경 (고급 저장 옵션) / 일부 캐스팅 경고 안 나오도록 static_cast<T> (특히 hlsl 과 int - float 축소 변환이 안 되어 타일이 렌더링 안 되던 문제 수정)

+++ 2DFrameWork/DWrite.cpp
@@ -170,6 +170,6 @@ void DWrite::RenderText(wstring text, RECT rect, float size, wstring font, Color
 
     deviceContext->DrawTextW
     (
-        text.c_str(), text.size(), findText->format, rectF, findBrush->brush
+        text.c_str(), (UINT32)text.size(), findText->format, rectF, findBrush->brush
     );
 }
